organizze external services - contains gprofiler and pathway commons - try to make the code less convoluted
