--- src/server/enrichment/analysis/index.js
@@ -1,4 +1,0 @@
-const { enrichment } = require('./enrichment');
-
-
-module.exports = { enrichment };
