--- src/server/enrichment/helper/index.js
@@ -1,4 +1,0 @@
-const { cleanUpEntrez } = require('./clean-up-entrez');
-
-
-module.exports = { cleanUpEntrez };
