--- src/server/enrichment/validation/validity-info.js
@@ -1,261 +1,0 @@
-// all valid organisms
-const validOrganism = ['aaegypti',
-  'acarolinensis',
-  'agambiae',
-  'amelanoleuca',
-  'amexicanus',
-  'aplatyrhynchos',
-  'btaurus',
-  'caperea',
-  'cchok1gshd',
-  'ccrigri',
-  'celegans',
-  'cfamiliaris',
-  'choffmanni',
-  'cintestinalis',
-  'cjacchus',
-  'clanigera',
-  'cporcellus',
-  'csabaeus',
-  'csavignyi',
-  'csyrichta',
-  'dmelanogaster',
-  'dnovemcinctus',
-  'dordii',
-  'drerio',
-  'ecaballus',
-  'eeuropaeus',
-  'etelfairi',
-  'falbicollis',
-  'fcatus',
-  'fdamarensis',
-  'gaculeatus',
-  'ggallus',
-  'ggorilla',
-  'gmorhua',
-  'hfemale',
-  'hmale',
-  'hsapiens',
-  'itridecemlineatus',
-  'jjaculus',
-  'lafricana',
-  'lchalumnae',
-  'loculatus',
-  'mauratus',
-  'mcaroli',
-  'mdomestica',
-  'mfuro',
-  'mgallopavo',
-  'mlucifugus',
-  'mmulatta',
-  'mmurinus',
-  'mmusculus',
-  'mochrogaster',
-  'mpahari',
-  'mspreteij',
-  'neugenii',
-  'ngalili',
-  'nleucogenys',
-  'oanatinus',
-  'oaries',
-  'ocuniculus',
-  'odegus',
-  'ogarnettii',
-  'olatipes',
-  'oniloticus',
-  'oprinceps',
-  'pabelii',
-  'panubis',
-  'pbairdii',
-  'pcapensis',
-  'pformosa',
-  'pmarinus',
-  'psinensis',
-  'ptroglodytes',
-  'pvampyrus',
-  'rnorvegicus',
-  'saraneus',
-  'scerevisiae',
-  'sharrisii',
-  'sscrofa',
-  'tbelangeri',
-  'tguttata',
-  'tnigroviridis',
-  'trubripes',
-  'ttruncatus',
-  'vpacos',
-  'xmaculatus',
-  'xtropicalis',
-  'aclavatus',
-  'aflavus',
-  'afumigatus',
-  'afumigatusa1163',
-  'agossypii',
-  'anidulans',
-  'aniger',
-  'aoryzae',
-  'aterreus',
-  'bbassiana',
-  'bcinerea',
-  'bgraminis',
-  'cgloeosporioides',
-  'cgraminicola',
-  'chigginsianum',
-  'cneoformans',
-  'corbiculare',
-  'dseptosporum',
-  'fculmorum',
-  'ffujikuroi',
-  'fgraminearum',
-  'foxysporum',
-  'fpseudograminearum',
-  'fsolani',
-  'fverticillioides',
-  'ggraminis',
-  'kpastoris',
-  'lmaculans',
-  'mlaricipopulina',
-  'moryzae',
-  'mpoae',
-  'mviolaceum',
-  'ncrassa',
-  'nfischeri',
-  'pgraminis',
-  'pgraminisug99',
-  'pnodorum',
-  'pstriiformis',
-  'pteres',
-  'ptriticina',
-  'ptriticirepentis',
-  'scerevisiae',
-  'scryophilus',
-  'sjaponicus',
-  'soctosporus',
-  'spombe',
-  'sreilianum',
-  'ssclerotiorum',
-  'tmelanosporum',
-  'treesei',
-  'tvirens',
-  'umaydis',
-  'vdahliae',
-  'ylipolytica',
-  'ztritici',
-  'aaegypti',
-  'acephalotes',
-  'adarlingi',
-  'agambiae',
-  'aglabripennis',
-  'amellifera',
-  'apisum',
-  'aqueenslandica',
-  'avaga',
-  'bantarctica',
-  'bimpatiens',
-  'bmalayi',
-  'bmori',
-  'cbrenneri',
-  'cbriggsae',
-  'celegans',
-  'cgigas',
-  'cjaponica',
-  'cquinquefasciatus',
-  'cremanei',
-  'cteleta',
-  'dananassae',
-  'derecta',
-  'dgrimshawi',
-  'dmelanogaster',
-  'dmojavensis',
-  'dpersimilis',
-  'dplexippus',
-  'dponderosae',
-  'dpseudoobscura',
-  'dpulex',
-  'dsechellia',
-  'dsimulans',
-  'dvirilis',
-  'dwillistoni',
-  'dyakuba',
-  'hmelpomene',
-  'hrobusta',
-  'iscapularis',
-  'lanatina',
-  'lcuprina',
-  'lgigantea',
-  'lloa',
-  'lsalmonis',
-  'mcinxia',
-  'mdestructor',
-  'mleidyi',
-  'mscalaris',
-  'nvectensis',
-  'nvitripennis',
-  'obimaculoides',
-  'ovolvulus',
-  'phumanus',
-  'ppacificus',
-  'rprolixus',
-  'sinvicta',
-  'smansoni',
-  'smaritima',
-  'smimosarum',
-  'spurpuratus',
-  'sratti',
-  'sscabiei',
-  'tadhaerens',
-  'tcastaneum',
-  'tkitauei',
-  'tspiralis',
-  'turticae',
-  'znevadensis',
-  'alyrata',
-  'atauschii',
-  'athaliana',
-  'atrichopoda',
-  'bdistachyon',
-  'bnapus',
-  'boleracea',
-  'brapa',
-  'bvulgaris',
-  'ccrispus',
-  'cmerolae',
-  'creinhardtii',
-  'gmax',
-  'gsulphuraria',
-  'hvulgare',
-  'lperrieri',
-  'macuminata',
-  'mtruncatula',
-  'obarthii',
-  'obrachyantha',
-  'oglaberrima',
-  'oglumaepatula',
-  'oindica',
-  'olongistaminata',
-  'olucimarinus',
-  'omeridionalis',
-  'onivara',
-  'opunctata',
-  'orufipogon',
-  'osativa',
-  'ppatens',
-  'ppersica',
-  'ptrichocarpa',
-  'sbicolor',
-  'sitalica',
-  'slycopersicum',
-  'smoellendorffii',
-  'stuberosum',
-  'taestivum',
-  'tcacao',
-  'tpratense',
-  'turartu',
-  'vvinifera',
-  'zmays'];
-
-// all valid target databases
-const validTargetDb = ['ENSG', 'HGNCSYMBOL', 'HGNC', 'UNIPROT', 'NCBIGENE'];
-
-
-module.exports = { validOrganism, validTargetDb };
