--- src/server/enrichment/visualization/index.js
@@ -1,4 +1,0 @@
-const { generateGraphInfo } = require('./generate-graph-info');
-
-
-module.exports = { generateGraphInfo };
